Move postCollection functions to related file
