+++ src/lib/filter/adjudicators/bindingAdjudicators.test.ts
@@ -5,7 +5,6 @@ import { expect, describe, it } from "@jest/globals";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import { Binding, DeepPartial, ValidateActionResponse } from "../../types";
 import { Event } from "../../enums";
-import { bindsToNamespace, misboundDeleteWithDeletionTimestamp, misboundNamespace } from "./postCollection";
 import { defaultBinding, defaultFilters, defaultKubernetesObject } from "./defaultTestObjects";
 import {
   definesDeletionTimestamp,
@@ -214,26 +213,6 @@ describe("missingName", () => {
   });
 });
 
-describe("bindsToNamespace", () => {
-  //[ Binding, result ]
-  it.each([
-    [{ kind: { kind: "" } }, false],
-    [{ kind: { kind: "Namespace" } }, true],
-  ])("given binding %j returns %s", (given, expected) => {
-    const binding: Binding = {
-      ...defaultBinding,
-      filters: {
-        ...defaultFilters,
-      },
-      kind: { kind: given.kind.kind, group: defaultBinding.kind.group },
-    };
-
-    const result = bindsToNamespace(binding);
-
-    expect(result).toBe(expected);
-  });
-});
-
 describe("definedNamespaces", () => {
   //[ Binding, result ]
   it.each([
@@ -347,29 +326,6 @@ describe("carriesNamespace", () => {
     };
 
     const result = carriesNamespace(kubernetesObject);
-
-    expect(result).toBe(expected);
-  });
-});
-
-describe("misboundNamespace", () => {
-  //[ Binding, result ]
-  it.each([
-    [{ kind: { kind: "Kind" }, filters: { namespaces: [] } }, false],
-    [{ kind: { kind: "Kind" }, filters: { namespaces: ["namespace"] } }, false],
-    [{ kind: { kind: "Namespace" }, filters: { namespaces: [] } }, false],
-    [{ kind: { kind: "Namespace" }, filters: { namespaces: ["namespace"] } }, true],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding: Binding = {
-      ...defaultBinding,
-      filters: {
-        ...defaultFilters,
-        namespaces: given.filters.namespaces,
-      },
-      kind: { kind: given.kind.kind, group: defaultBinding.kind.group },
-    };
-
-    const result = misboundNamespace(binding);
 
     expect(result).toBe(expected);
   });
@@ -564,41 +520,6 @@ describe("definesDelete", () => {
     };
 
     const result = definesDelete(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-
-describe("misboundDeleteWithDeletionTimestamp", () => {
-  //[ Binding, result ]
-  it.each([
-    [{ event: Event.DELETE, filters: { deletionTimestamp: false } }, false],
-    [{ event: Event.DELETE, filters: { deletionTimestamp: true } }, true],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding: Binding = {
-      ...defaultBinding,
-      filters: { ...defaultFilters, deletionTimestamp: given.filters.deletionTimestamp },
-      event: given.event,
-    };
-
-    const result = misboundDeleteWithDeletionTimestamp(binding);
-
-    expect(result).toEqual(expected);
-  });
-});
-describe("when filters are not set", () => {
-  it.each([
-    [{ event: Event.CREATE }, false],
-    [{ event: Event.CREATE_OR_UPDATE }, false],
-    [{ event: Event.UPDATE }, false],
-    [{ event: Event.DELETE }, false],
-  ])("given %j, returns %s", (given, expected) => {
-    const binding: Binding = {
-      ...defaultBinding,
-      event: given.event,
-    };
-
-    const result = misboundDeleteWithDeletionTimestamp(binding);
 
     expect(result).toEqual(expected);
   });
