+++ src/lib/filter/adjudicators/postCollection.test.ts
@@ -4,15 +4,19 @@
 import { expect, describe, it } from "@jest/globals";
 import {
   bindsToKind,
+  bindsToNamespace,
   carriesIgnoredNamespace,
+  misboundDeleteWithDeletionTimestamp,
+  misboundNamespace,
   missingCarriableNamespace,
-  operationMatchesEvent,
   unbindableNamespaces,
   uncarryableNamespace,
 } from "./postCollection";
+import { operationMatchesEvent } from "./mismatch";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import { Binding, DeepPartial } from "../../types";
 import { Event, Operation } from "../../enums";
+import { defaultBinding, defaultFilters } from "./defaultTestObjects";
 
 describe("bindsToKind", () => {
   //[ Binding, Kind, result ]
@@ -189,6 +193,84 @@ describe("operationMatchesEvent", () => {
     [Operation.CONNECT, Event.ANY, true],
   ])("given operation %s and event %s, returns %s", (op, evt, expected) => {
     const result = operationMatchesEvent(op, evt);
+
+    expect(result).toEqual(expected);
+  });
+});
+
+describe("bindsToNamespace", () => {
+  //[ Binding, result ]
+  it.each([
+    [{ kind: { kind: "" } }, false],
+    [{ kind: { kind: "Namespace" } }, true],
+  ])("given binding %j returns %s", (given, expected) => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: {
+        ...defaultFilters,
+      },
+      kind: { kind: given.kind.kind, group: defaultBinding.kind.group },
+    };
+
+    const result = bindsToNamespace(binding);
+
+    expect(result).toBe(expected);
+  });
+});
+
+describe("misboundNamespace", () => {
+  //[ Binding, result ]
+  it.each([
+    [{ kind: { kind: "Kind" }, filters: { namespaces: [] } }, false],
+    [{ kind: { kind: "Kind" }, filters: { namespaces: ["namespace"] } }, false],
+    [{ kind: { kind: "Namespace" }, filters: { namespaces: [] } }, false],
+    [{ kind: { kind: "Namespace" }, filters: { namespaces: ["namespace"] } }, true],
+  ])("given %j, returns %s", (given, expected) => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: {
+        ...defaultFilters,
+        namespaces: given.filters.namespaces,
+      },
+      kind: { kind: given.kind.kind, group: defaultBinding.kind.group },
+    };
+
+    const result = misboundNamespace(binding);
+
+    expect(result).toBe(expected);
+  });
+});
+
+describe("misboundDeleteWithDeletionTimestamp", () => {
+  //[ Binding, result ]
+  it.each([
+    [{ event: Event.DELETE, filters: { deletionTimestamp: false } }, false],
+    [{ event: Event.DELETE, filters: { deletionTimestamp: true } }, true],
+  ])("given %j, returns %s", (given, expected) => {
+    const binding: Binding = {
+      ...defaultBinding,
+      filters: { ...defaultFilters, deletionTimestamp: given.filters.deletionTimestamp },
+      event: given.event,
+    };
+
+    const result = misboundDeleteWithDeletionTimestamp(binding);
+
+    expect(result).toEqual(expected);
+  });
+});
+describe("when filters are not set", () => {
+  it.each([
+    [{ event: Event.CREATE }, false],
+    [{ event: Event.CREATE_OR_UPDATE }, false],
+    [{ event: Event.UPDATE }, false],
+    [{ event: Event.DELETE }, false],
+  ])("given %j, returns %s", (given, expected) => {
+    const binding: Binding = {
+      ...defaultBinding,
+      event: given.event,
+    };
+
+    const result = misboundDeleteWithDeletionTimestamp(binding);
 
     expect(result).toEqual(expected);
   });
